Revert "Refactor Focus Event Ping to use SendsToAmplitude trait"
This reverts commit 0b364e1303fb108a5b3b242f2d8c4d2c801fece0.
