+++ src/main/scala/com/mozilla/telemetry/pings/FocusEvent.scala
@@ -3,11 +3,16 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 package com.mozilla.telemetry.pings
 
+import java.security.MessageDigest
+
+import com.fasterxml.jackson.databind.JsonNode
+import com.github.fge.jsonschema.main.JsonSchemaFactory
 import com.mozilla.telemetry.heka.Message
 import com.mozilla.telemetry.pings.Ping.messageToPing
-import com.mozilla.telemetry.streaming.EventsToAmplitude.{AmplitudeEvent, Config}
+import com.mozilla.telemetry.streaming.EventsToAmplitude.Config
 import org.json4s.JsonDSL._
 import org.json4s._
+import org.json4s.jackson.JsonMethods._
 
 case class FocusEventPing(clientId: String,
                           created: Long,
@@ -17,33 +22,92 @@ case class FocusEventPing(clientId: String,
                           os: String,
                           osversion: String,
                           settings: FocusSettings,
-                          meta: Meta) extends SendsToAmplitude {
-  def sessionStart: Long = created
+                          meta: Meta) {
 
-  def getSessionId: String = (events.map(_.timestamp).max).toString
+  def getEvents(config: Config): String = {
+    implicit val formats = DefaultFormats
 
-  def getOs: String = os
+    val factory = JsonSchemaFactory.byDefault
+    val schemas = config.events.map(e => factory.getJsonSchema(asJsonNode(e.schema)))
 
-  def getOsVersion: String = osversion
+    val clientid = getHashedClientId
+    val sessionId = getSessionId
 
-  override def eventToAmplitudeEvent(config: Config, e: Event, es: AmplitudeEvent): JObject = {
-    super.eventToAmplitudeEvent(config, e, es) ~
-      ("user_properties" ->
-        ("pref_privacy_block_ads" -> settings.blockAds) ~
-          ("pref_locale" -> settings.pref_locale) ~
-          ("pref_privacy_block_social" -> settings.blockSocial) ~
-          ("pref_secure" -> settings.secure) ~
-          ("pref_privacy_block_analytics" -> settings.blockAnalytics) ~
-          ("pref_search_engine" -> settings.pref_search_engine) ~
-          ("pref_privacy_block_other" -> settings.blockOther) ~
-          ("pref_default_browser" -> settings.defaultBrowser) ~
-          ("pref_performance_block_webfonts" -> settings.blockWebfonts) ~
-          ("pref_performance_block_images" -> settings.blockImages) ~
-          ("pref_autocomplete_installed" -> settings.autocompleteInstalled) ~
-          ("pref_autocomplete_custom" -> settings.autocompleteCustom))
+    val eventsList = events.map { e => e -> asJsonNode(Extraction.decompose(e)): (Event, JsonNode) }
+      .map { case (e, es) => // for each event, try each schema
+        e -> schemas.map(ts => ts.validateUnchecked(es).isSuccess)
+          .zip(config.events)
+          .filter(_._1)
+      }
+      .filter { case (e, es) => !es.isEmpty } // only keep those with a match
+      .map { case (e, es) => e -> es.head._2 } // take the first match (head._1 is the bool)
+      .map { case (e, es) =>
+      ("device_id" -> clientId) ~
+        ("session_id" -> sessionId) ~
+        ("insert_id" -> (clientId + sessionId.toString + e.getId)) ~
+        ("event_type" -> getFullEventName(config.eventGroupName, es.name)) ~
+        ("time" -> (e.timestamp + created)) ~
+        ("event_properties" -> e.getProperties(es.amplitudeProperties)) ~
+        ("app_version" -> meta.appVersion) ~
+        ("os_name" -> os) ~
+        ("os_version" -> osversion) ~
+        ("country" -> meta.geoCountry) ~
+        ("city" -> meta.geoCity) ~
+        ("user_properties" ->
+          ("pref_privacy_block_ads" -> settings.blockAds) ~
+            ("pref_locale" -> settings.pref_locale) ~
+            ("pref_privacy_block_social" -> settings.blockSocial) ~
+            ("pref_secure" -> settings.secure) ~
+            ("pref_privacy_block_analytics" -> settings.blockAnalytics) ~
+            ("pref_search_engine" -> settings.pref_search_engine) ~
+            ("pref_privacy_block_other" -> settings.blockOther) ~
+            ("pref_default_browser" -> settings.defaultBrowser) ~
+            ("pref_performance_block_webfonts" -> settings.blockWebfonts) ~
+            ("pref_performance_block_images" -> settings.blockImages) ~
+            ("pref_autocomplete_installed" -> settings.autocompleteInstalled) ~
+            ("pref_autocomplete_custom" -> settings.autocompleteCustom))
+    }
+
+    compact(render(eventsList))
+  }
+
+  def getFullEventName(groupName: String, eventName: String): String = groupName + " - " + eventName
+
+  def getSessionId: Long = events.map(_.timestamp).max
+
+  def getHashedClientId: String = {
+    val messageDigest = MessageDigest.getInstance("SHA-256")
+    new String(messageDigest.digest(clientId.getBytes()))
+  }
+
+  def includePing(sample: Double, config: Config): Boolean = {
+    val keepClient = meta.sampleId.getOrElse(sample * 100) < (sample * 100)
+
+    if (!keepClient) {
+      // for maybe a slight perf increase
+      return false // scalastyle:ignore
+    }
+
+    val currentProps = this.getClass
+      .getDeclaredFields.map { e =>
+      e.setAccessible(true)
+      e.getName -> e.get(this).toString
+    }.toMap
+
+    config.filters.map { case (prop, allowedVals) =>
+      allowedVals.contains(currentProps.getOrElse(prop, allowedVals.head))
+    }.foldLeft(true)(_ & _)
   }
 }
 
+object FocusEventPing {
+  def apply(message: Message): FocusEventPing = {
+    implicit val formats = DefaultFormats
+    val jsonFieldNames = List()
+    val ping = messageToPing(message, jsonFieldNames, List("events" :: Nil))
+    ping.extract[FocusEventPing]
+  }
+}
 
 case class Event(timestamp: Int,
                  category: String,
